Make Unbound UBsan aware

+++ build-unbound.sh
@@ -97,6 +97,14 @@ then
     [[ "$0" = "${BASH_SOURCE[0]}" ]] && exit 1 || return 1
 fi
 
+echo "Searching for errors hidden in log files"
+COUNT=$(grep -oIR 'runtime error' | wc -l)
+if [[ "${COUNT}" -ne 0 ]];
+then
+    echo "Failed to test Unbound"
+    [[ "$0" = "${BASH_SOURCE[0]}" ]] && exit 1 || return 1
+fi
+
 rm -f root.key 2>/dev/null
 UNBOUND_ANCHOR_PROG=$(find "$PWD" -name unbound-anchor | head -n 1)
 if [[ -z "$UNBOUND_ANCHOR_PROG" ]]; then
